Moving to sway

+++ custom/user_scripts/diff-so-fancy.sh
@@ -5,4 +5,6 @@ mkdir -p ~/bin
 cd ~/bin
 git clone https://github.com/so-fancy/diff-so-fancy diffsofancy
 chmod +x diffsofancy/diff-so-fancy
-ln -s ~/bin/diffsofancy/diff-so-fancy ~/bin/diff-so-fancy
+if [ ! -f ~/bin/diff-so-fancy ] ; then
+	ln -s ~/bin/diffsofancy/diff-so-fancy ~/bin/diff-so-fancy 
+fi
